+++ packages/ActionBar/src/hooks/useColumnsArrangement/useColumnsArrangement.js
@@ -1,18 +1,29 @@
 import React from "react";
 import produce from "immer";
 
-export default function useColumnsArrangement(defaultOrder) {
+export default function useColumnsArrangement(defaultOrder, disabledColumnIds = []) {
   const [order, setOrder] = React.useState(defaultOrder);
-  const [hiddenColumnIds, setHiddenColumnIds] = React.useState([]);
+  const [hiddenColumnIds, setHiddenColumnIds] = React.useState(new Set());
 
-  const isColumnHidden = React.useCallback(
-    columnId => {
-      return hiddenColumnIds.indexOf(columnId) > -1;
-    },
-    [hiddenColumnIds]
-  );
+  function canMove({ source, destination }) {
+    if (disabledColumnIds.length === 0) return true;
+
+    const newOrder = [...order];
+    const movedChild = newOrder.splice(source, 1);
+    newOrder.splice(destination, 0, ...movedChild);
+
+    return disabledColumnIds.every(
+      disabledColumnId => defaultOrder.indexOf(disabledColumnId) === newOrder.indexOf(disabledColumnId)
+    );
+  }
+
+  function isColumnHidden(columnId) {
+    return hiddenColumnIds.has(columnId);
+  }
 
   function handleChangeOrder({ source, destination }) {
+    if (!canMove({ source, destination })) return;
+
     setOrder(
       produce(draftOrder => {
         const movedChild = draftOrder.splice(source, 1);
@@ -22,22 +33,22 @@ export default function useColumnsArrangement(defaultOrder) {
   }
 
   function handleHideAll() {
-    setHiddenColumnIds([...order]);
+    setHiddenColumnIds(new Set(order.filter(id => !disabledColumnIds.includes(id))));
   }
 
   function handleShowAll() {
-    setHiddenColumnIds([]);
+    setHiddenColumnIds(new Set());
   }
 
   function handleChangeVisibility(columnId) {
     setHiddenColumnIds(
       produce(draft => {
-        const index = draft.indexOf(columnId);
-        if (index > -1) {
-          draft.splice(index, 1);
+        if (draft.has(columnId)) {
+          draft.delete(columnId);
         } else {
-          draft.push(columnId);
+          draft.add(columnId);
         }
+        return new Set(draft);
       })
     );
   }