+++ packages/Panel/src/index.d.ts
@@ -11,27 +11,29 @@ interface PanelProps {
   getPushContentElement?: (...args: any[]) => any;
   /** Y offset that is passed down from <Panel.Group> */
   groupOffsetY?: number;
-  /** Control the compactness of the side panel */
+  /** The height of the open Panel (when slide in from bottom) */
+  height?: string | number;
+  /** Control the compactness of the Panel */
   isCompact?: boolean;
   /** Render the panel inline */
   isInline?: boolean;
-  /** Control the visibility of the side panel. This prop makes the side panel appear */
+  /** Control the visibility of the Panel. This prop makes the Panel appear. */
   isOpen: boolean;
-  /** Control if the side panel slides from the left */
-  isSlideFromLeft?: boolean;
   /** Modify the look of the Panel */
   kind?: Panel.types.kind.DEFAULT | Panel.types.kind.CHILD | Panel.types.kind.PRIMARY;
-  /** Control y offset of the panel */
-  offsetY?: number;
-  /** Callback once the panel has been closed event */
+  /** Control offset of the Panel. Only use 'top' when sliding in from the left or right. Only use 'left' or 'right' when sliding in from the bottom. */
+  offset?: shape;
+  /** Callback once the Panel has been closed event */
   onAfterClose?: (...args: any[]) => any;
-  /** Callback once the panel has been opened event */
+  /** Callback once the Panel has been opened event */
   onAfterOpen?: (...args: any[]) => any;
-  /** Callback triggered when the side panel needs to be close */
+  /** Callback triggered when the Panel needs to be close */
   onClose?: (...args: any[]) => any;
-  /** The width of the open panel. */
+  /** Control where the Panel slides in from */
+  slideFrom?: Panel.types.slideFroms.RIGHT | Panel.types.slideFroms.LEFT | Panel.types.slideFroms.BOTTOM;
+  /** The width of the open Panel (when slide in from left or right) */
   width?: string | number;
-  /** Control the z position of the panel */
+  /** Control the z-index of the Panel */
   zIndex?: number;
 }
 
@@ -41,6 +43,15 @@ declare namespace Panel {
       const DEFAULT: any;
       const CHILD: any;
       const PRIMARY: any;
+    }
+  }
+}
+declare namespace Panel {
+  namespace types {
+    namespace slideFroms {
+      const RIGHT: any;
+      const LEFT: any;
+      const BOTTOM: any;
     }
   }
 }