+++ packages/Panel/stories/examples/Examples.js
@@ -7,7 +7,7 @@ import Panel from "../../src";
 import { Nav, TextLine } from "../helpers";
 
 export const BasicPanel = props => {
-  const { disableBodyOverflow, hasOverlay, isSlideFromLeft } = props;
+  const { disableBodyOverflow, hasOverlay, slideFrom } = props;
   const [isOpen, setIsOpen] = React.useState(true);
   const toggle = () => {
     setIsOpen(state => !state);
@@ -21,8 +21,8 @@ export const BasicPanel = props => {
         disableBodyOverflow={disableBodyOverflow}
         isOpen={isOpen}
         onClose={toggle}
-        offsetY={40}
-        isSlideFromLeft={isSlideFromLeft}
+        offset={{ top: 40 }}
+        slideFrom={slideFrom}
       >
         {hasOverlay ? <Panel.Overlay /> : null}
         <Panel.Trigger kind="primary" onClick={toggle}>
@@ -50,6 +50,55 @@ export const FocusLockPanel = () => {
         <input type="text" data-autofocus />
       </Panel.Content>
     </Panel>
+  );
+};
+
+export const FromBottom = () => {
+  const [isFooterOpen, setIsFooterOpen] = React.useState(true);
+  const [isNavOpen, setIsNavOpen] = React.useState(true);
+  const navWidth = { expanded: 300, collapsed: 30 };
+
+  const toggleFooter = () => {
+    setIsFooterOpen(state => !state);
+  };
+  const toggleNav = () => {
+    setIsNavOpen(state => !state);
+  };
+
+  function getNavWidth() {
+    return isNavOpen ? navWidth.expanded : navWidth.collapsed;
+  }
+
+  return (
+    <React.Fragment>
+      <div style={{ display: "flex", height: "100vh" }}>
+        <nav style={{ flex: `0 0 ${getNavWidth()}px`, background: "red" }}>
+          <button type="button" onClick={toggleNav}>
+            {isNavOpen ? <span>&laquo;</span> : <span>&raquo;</span>}
+          </button>
+        </nav>
+        <div style={{ flex: 1, padding: "10px" }}>
+          <p>Try toggling the div on the left and notice the Panel&apos;s width adjusts (controlled by consumer).</p>
+          <button type="button" onClick={toggleFooter}>
+            {isFooterOpen ? "close" : "open"} the bottom Panel
+          </button>
+        </div>
+      </div>
+
+      <Panel
+        height={150}
+        isOpen={isFooterOpen}
+        offset={{ left: getNavWidth() }}
+        onClose={toggleFooter}
+        slideFrom={Panel.types.slideFrom.BOTTOM}
+      >
+        <Panel.Content>
+          This Panel comes up from the bottom.
+          <TextLine repeat={10} />
+          Pretty neat, eh?
+        </Panel.Content>
+      </Panel>
+    </React.Fragment>
   );
 };