+++ packages/Sortable/CHANGELOG.md
@@ -5,3 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [1.0.2-alpha.2] - 2020-11-24
+
+- Added `isDragDisabled` and `handleElement` prop on `Sortable.Item`.